**** Code is Cleaner! *****
I have commented all the code in the html so that it is clear. Try to
follow my style of commenting. Usually, I write a comment for large
sections of the page (header, content, footer, etc.), sections of the
page that may not be as clear, and scripts. Also, be sure to use
template.html to build new pages